fix(api): relocate git_context into scripts/common (root-scripts gate, 2nd file)
The gate's violation output names scripts/git_context.py — added at
scripts/ root by this PR's original build (release_layout was only the
first of two). Moved to scripts/common/ alongside release_layout; all
six importers updated (worktree_containment, decision_lineage,
git_hygiene_router, worktrees_router, release_snapshot, tests).

X-Agent: claude-infra
