+++ scripts/api/git_hygiene_router.py
@@ -30,7 +30,7 @@ from typing import Any
 from fastapi import APIRouter
 from pydantic import BaseModel
 
-from scripts.git_context import sanitized_git_env
+from scripts.common.git_context import sanitized_git_env
 
 from .config import LIVE_REPO_ROOT, PROJECT_ROOT
