+++ scripts/common/git_context.py
