+++ tests/api/test_release_snapshot.py
@@ -16,8 +16,8 @@ from pathlib import Path
 import pytest
 
 from scripts.api import release_snapshot
+from scripts.common.git_context import sanitized_git_env
 from scripts.common.release_layout import MANIFEST_NAME, is_release_root
-from scripts.git_context import sanitized_git_env
 from scripts.path_safety import safe_join
 
 PROJECT_ROOT = Path(__file__).resolve().parents[2]
